Reverted patch from David Chisnall to bring back Lucid to a working state.

+++ sope-appserver/NGObjWeb/WebDAV/SaxDAVHandler.m
@@ -928,7 +928,7 @@ static BOOL heavyLog = NO;
 
 /* CDATA */
 
-- (void)characters:(unichar *)_chars length:(NSUInteger)_len {
+- (void)characters:(unichar *)_chars length:(int)_len {
   if (heavyLog) [self logWithFormat:@"got %i chars", _len];
   
   if (_len > 0 && (self->cdata != nil)) {
