+++ sope-appserver/WEExtensions/WEResourceKey.m
@@ -65,7 +65,7 @@
 
 /* equality */
 
-- (NSUInteger)hash {
+- (unsigned)hash {
   if (self->hashValue == 0) {
     /* don't know whether this is smart, Nat! needs to comment ;-) */
     self->hashValue = [self->name hash];
